Move the `file-store` module under the `@bayou` namespace.
