+++ doc/coding-conventions.md
@@ -95,7 +95,7 @@ taking into account recent additions to the language.
   import express from 'express';
   import fs from 'fs';
 
-  import { BaseFile } from 'file-store';
+  import { BaseFile } from '@bayou/file-store';
   import { DataUtil, InfoError, Singleton } from '@bayou/util-common';
 
   import RegularBlort from './RegularBlort';
@@ -129,8 +129,8 @@ taking into account recent additions to the language.
   methods together within a class as being part of a "namespace" of some sort.
   When this is the case, the names take the form `<namespace>_<method>` where
   the namespace and method name are underscore-separated. For example, the
-  transaction operation constructors in `file-store.TransactionOp` are all defined
-  as `op_<name>`.
+  transaction operation constructors in `@bayou/file-store.TransactionOp` are
+  all defined as `op_<name>`.
 
 * Abstract methods &mdash; Sometimes base classes want to define abstract
   methods for their subclasses to fill in. The naming pattern for these is a
