+++ local-modules/@bayou/file-store/mocks/MockFile.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { BaseFile } from 'file-store';
+import { BaseFile } from '@bayou/file-store';
 
 /**
  * Trivial {@link BaseFile} implementation for use with the tests.
