+++ local-modules/@bayou/file-store/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "file-store",
+  "name": "@bayou/file-store",
   "version": "1.0.0",
 
   "dependencies": {
