+++ local-modules/@bayou/hooks-server/Hooks.js
@@ -39,7 +39,7 @@ export default class Hooks extends Singleton {
    * {BaseFileStore} The object which provides access to file storage (roughly
    * speaking, the filesystem to store the "files" this system deals with). This
    * is an instance of a subclass of `BaseFileStore`, as defined by the
-   * `file-store` module.
+   * `@bayou/file-store` module.
    */
   get fileStore() {
     return LocalFileStore.theOne;
