+++ local-modules/doc-server/package.json
@@ -9,7 +9,7 @@
     "@bayou/codec": "local",
     "@bayou/doc-common": "local",
     "@bayou/env-server": "local",
-    "file-store": "local",
+    "@bayou/file-store": "local",
     "@bayou/file-store-ot": "local",
     "@bayou/hooks-server": "local",
     "@bayou/ot-common": "local",
