+++ local-modules/file-store-local/package.json
@@ -6,7 +6,7 @@
     "@bayou/codec": "local",
     "@bayou/deps-util-server": "local",
     "@bayou/env-server": "local",
-    "file-store": "local",
+    "@bayou/file-store": "local",
     "@bayou/file-store-ot": "local",
     "@bayou/promise-util": "local",
     "@bayou/see-all": "local",
